Separate the hubcap package from the hubcap.py script for testability
